add back booking cancellation own user verification this time working

+++ src/graphql/resolvers/booking.ts
@@ -44,6 +44,9 @@ module.exports = {
     }
     try {
       const booking = await Booking.findById(args.bookingId).populate('event')
+      if (req.userId.toString() !== booking.user.toString()) {
+        throw new Error('User not deleting its own booking')
+      }
       const event: IEventDocument['_doc'] = standarizeEvent(booking.event)
       await Booking.deleteOne({ _id: args.bookingId })
       return event
